pyproject.toml
[tool.poetry]
name = "the-fig-tree"
version = "1.0.0"
description = "The Fig Tree is my corner of the world to write and enjoy theehe fruits of my cognitive experiences in our culture."
authors = ["John Paul Jayme"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
markdown = "^3.7"
python-frontmatter = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
